fix: setup page to open setup screen
