+++ src/renderer/src/assets/store.css
@@ -36,5 +36,10 @@
   width: 100%;
   height: 50px;
   text-transform: uppercase;
-  margin-top: 15px
+  margin-top: 15px;
+  cursor: pointer;
+}
+
+.store-button:hover {
+  background: #49677b;
 }
